I3 dock fix (#258)
* fix: problem with docked widgets on i3

Show window only after settings x-props. According to the EWMH spec:
> _NET_WM_WINDOW_TYPE
> This SHOULD be set by the Client before mapping to a list of atoms
> indicating the functional type of the window.

So the _NET_WM_WINDOW_TYPE atom must be set before showing the window.

The callback to the configure event handle now checks the windows
current geometry to not move the window if it's already where it's
supposed to be. This prevents the eww daemon from endlessly chatting
with X11 in some instances.

* chore: cargo fmt