Remove window flicker from open command
This solves a problem that had been already solved in
0d1b5e0f9e532ade4b7c9fd77b46b6a29b1d8529 but was reintroduced in
a9a35c1804d72ef92e04ee71555bd9e5a08fa17e. Basically before this, the
window has been closed and reopened if it existed, which results in
flickering due to the animations.

In addition to remove the window flicker the timer (if started with
--duration) will also be renewed. This fixes #892.