Log to file also when not started from terminal
To capture logs into a file we redirect stdout and stdin to a log file,
but this is currently only done if stdout or stdin are tty's. Whenever
eww is started from a background task, that is not directly from a
terminal, stdout and stdin aren't tty's and nothing is written to the
log files as a result.

I see no reason to require stdout and stdin to be ttys and I don't see
any other checks that would make sense instead, so just  remove the
isatty checks for stdout and stdin, and just always redirect them.